requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Django==5.2.8
django-admin-interface==0.31.0
django-allauth==65.13.1
django-axes==8.0.0
django-colorfield==0.14.0
django-extensions==4.1
django-preventconcurrentlogins==0.8.2
django-tailwind==4.4.1
et_xmlfile==2.0.0
fido2==2.0.0
gunicorn==23.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyMySQL==1.1.2
pytailwindcss==0.3.0
python-slugify==8.0.4
qrcode==8.2
reportlab==4.4.5
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
whitenoise==6.11.0
